+++ AugGISConfigTool/AugGISDataParser/source/LayerTypeData.cs
@@ -9,12 +9,12 @@ public class LayerTypeData
     public string polygonPatternName = "0";
     
     public string lineColor = "#000000";
-    public int lineWidth = 1;
+    public float lineWidth = 1;
     public string lineIcon = "";
     public string linePatternType = "Solid";
 
     public string pointColor = "#000000";
-    public int pointSize = 0;
+    public float pointSize = 0;
     public string pointSpriteName = "None";
 
     public string description = "";
